requirements.txt
requests>=2.28.0
genanki>=0.13.0
python-dotenv>=0.19.0
